Move DeviceSynchronization from pairing command to device_manager
