+++ examples/fabric-admin/BUILD.gn
@@ -76,14 +76,14 @@ static_library("fabric-admin-utils") {
     "commands/common/StayActiveSender.cpp",
     "commands/common/StayActiveSender.h",
     "commands/fabric-sync/FabricSyncCommand.cpp",
-    "commands/pairing/DeviceSynchronization.cpp",
-    "commands/pairing/DeviceSynchronization.h",
     "commands/pairing/OpenCommissioningWindowCommand.cpp",
     "commands/pairing/OpenCommissioningWindowCommand.h",
     "commands/pairing/PairingCommand.cpp",
     "commands/pairing/ToTLVCert.cpp",
     "device_manager/DeviceManager.cpp",
     "device_manager/DeviceManager.h",
+    "device_manager/DeviceSynchronization.cpp",
+    "device_manager/DeviceSynchronization.h",    
   ]
 
   deps = [ "${chip_root}/src/app:events" ]
