+++ examples/fabric-admin/device_manager/DeviceSynchronization.h
